python/requirements.txt
flask
SQLAlchemy
mysqlclient
mysql-connector-python
python-dotenv
redis
